:truck: rename: 카테고리,게임채널 페이지 관련 컴포넌트들 폴더 위치 이동 레이아웃에서 컴포넌트 하위폴더로 위치이동시킴
