+++ client/src/components/GameChannel/PostList.tsx
