+++ client/src/pages/CategoryGames.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import GameList from '../layouts/CategoryGames/GameList';
+import GameList from '../components/CategoryGames/GameList';
 import FilterTap from '../components/common/FilterTap';
-import RecommendGames from '../layouts/CategoryGames/RecommendGames';
-import TitleCategory from '../layouts/CategoryGames/TitleCategory';
+import RecommendGames from '../components/CategoryGames/RecommendGames';
+import TitleCategory from '../components/CategoryGames/TitleCategory';
 import { categoryFilterTab } from '../data/filterTapList';
 
 const CategoryGames = () => {
